I try github ui it is ok, and updateMotors v1

+++ src/main/java/com/winricklabs/mouse/dto/UpdateMotors.java
@@ -1,6 +1,6 @@
 package com.winricklabs.mouse.dto;
 
 public class UpdateMotors {
-    public double leftMotor;
-    public double rightMotor;
+    public int leftMotor;
+    public int rightMotor;
 }
